pyproject.toml
[tool.poetry]
name = "cli-chat"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
license = "LICENSE"
packages = [{include = "cli_chat"}]

[tool.poetry.dependencies]
python = "^3.10"
aioconsole = "0.4.1"
pycryptodome = "3.14.1"

[tool.poetry.group.dev.dependencies]
pyproject-flake8 = "6.0.0"
mypy = "0.800"

[tool.flake8]
exclude = ["venv", ".git", "__pycache__", ".mypy_cache"]
ignore = ["E402", "W503"]
max-line-length = 79

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
